Move classes from Gui Package to yapper package.
Attempt to fix "Could not find or load main class Launcher" issue when running Yapper.jar
