+++ src/main/java/yapper/Main.java
@@ -1,4 +1,4 @@
-package yapper.gui;
+package yapper;
 
 import java.io.IOException;
 
@@ -7,7 +7,6 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
-import yapper.Yapper;
 
 /**
  * A GUI for Duke using FXML.
