+++ src/main/java/yapper/MainWindow.java
@@ -1,4 +1,4 @@
-package yapper.gui;
+package yapper;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -7,7 +7,6 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-import yapper.Yapper;
 
 /**
  * Controller for the main GUI.
